agents/google_adk_agent.py
# agents/google_adk_agent.py
# Google ADK (Android Development Kit) Integration for Multi-Agent System
# This simulates Google ADK functionality for agent coordination

import os
from typing import Dict, List, Any
import google.generativeai as genai

class GoogleADKCoordinator:
    """
    Google ADK-style coordinator for managing agent workflows
    Uses Google's Gemini API for intelligent agent planning and coordination
    """
    def __init__(self):
        api_key = os.getenv("GOOGLE_API_KEY")
        if not api_key:
            raise ValueError("GOOGLE_API_KEY not found in environment variables")
        
        genai.configure(api_key=api_key)
        self.model = genai.GenerativeModel('gemini-1.5-flash')
    
    def plan_agent_sequence(self, user_goal: str) -> List[str]:
        """
        Use Google's Gemini to intelligently plan agent sequence
        """
        prompt = f"""
        You are an intelligent agent planner. Given a user goal, determine the optimal sequence of specialized agents to achieve it.
        
        Available agents:
        - spacex_agent: Gets SpaceX launch data and coordinates
        - weather_agent: Gets weather data for specific coordinates
        - summary_agent: Analyzes data and provides conclusions
        - news_agent: Gets relevant news articles (if needed)
        
        User Goal: "{user_goal}"
        
        Respond with ONLY a comma-separated list of agent names in execution order.
        Example: spacex_agent, weather_agent, summary_agent
        """
        
        try:
            response = self.model.generate_content(prompt)
            # Parse the response to extract agent names
            agent_list = [agent.strip() for agent in response.text.strip().split(',')]
            return agent_list
        except Exception as e:
            print(f"⚠️ ADK Planning fallback due to: {e}")            # Fallback to simple planning
            if "spacex" in user_goal.lower() and "weather" in user_goal.lower():
                return ["spacex_agent", "weather_agent", "summary_agent"]
            return []
    
    def validate_goal_completion(self, user_goal: str, final_data: Dict[str, Any]) -> Dict[str, Any]:
        """
        Use Google ADK to validate if the goal was achieved and suggest improvements
        """
        prompt = f"""
        Analyze if the user's goal was successfully achieved based on the data collected.
        
        User Goal: "{user_goal}"
        Final Data: {final_data}
        
        Provide a JSON response with:
        {{
            "goal_achieved": true/false,
            "confidence": 0-100,
            "missing_data": ["list of missing information"],
            "suggested_improvements": ["list of suggestions"],
            "quality_score": 0-100
        }}
        """
        
        try:
            response = self.model.generate_content(prompt)
            # Clean and parse JSON response
            json_text = response.text.strip()
            if json_text.startswith("```json"):
                json_text = json_text[7:-3]
            elif json_text.startswith("```"):
                json_text = json_text[3:-3]
            
            import json
            validation_result = json.loads(json_text)
            return validation_result
        except Exception as e:
            print(f"⚠️ ADK Validation fallback due to: {e}")
            return {
                "goal_achieved": True,
                "confidence": 80,
                "missing_data": [],
                "suggested_improvements": [],
                "quality_score": 80
            }

def run(previous_data: dict) -> dict:
    """
    Google ADK agent runner - this validates and improves the final result
    """
    adk = GoogleADKCoordinator()
    user_goal = previous_data.get("goal", "")
    
    validation = adk.validate_goal_completion(user_goal, previous_data)
    previous_data.update({"adk_validation": validation})
    
    return previous_data
