requirements.txt
requests
python-dotenv
google-adk
flask
google-generativeai